+++ test/specs/basic-x-2.js
@@ -17,6 +17,16 @@ document.write(
     '</div>'
 );
 
+function waitsFor (test, fn) {
+  if (test()) {
+    fn();
+  } else {
+    setTimeout(function () {
+      waitsFor(test, fn);
+    }, 10);
+  }
+}
+
 var fotorama1, fotorama2;
 
 function eachFotorama (callback) {
@@ -58,18 +68,18 @@ describe('Two fotoramas', function () {
       expect($('.fotorama__nav__frame', $fotorama).size()).toEqual(5 - i);
     });
   });
-  it('dimensions are correct', function () {
+  it('dimensions are correct', function (done) {
     eachFotorama(function (i, fotorama) {
       var $fotorama = $('#fotorama' + (i + 1)),
           $stage = $('.fotorama__stage', $fotorama);
 
       if (!i) {
         waitsFor(function () {
+          console.log('$stage.width()', $stage.width());
           return $stage.width() === 700
-        }, 'Waits for autosize...', 100);
-
-        runs(function () {
+        }, function () {
           expect($stage.height()).toEqual(467);
+          done();
         });
       } else {
         expect($stage.width()).toEqual(500);