+++ test/specs/basic.js
@@ -8,6 +8,16 @@ document.write(
     '</div>'
 );
 
+function waitsFor (test, fn) {
+  if (test()) {
+    fn();
+  } else {
+    setTimeout(function () {
+      waitsFor(test, fn);
+    }, 10);
+  }
+}
+
 describe('Basic', function () {
   var $fotorama, fotorama;
 
@@ -34,17 +44,20 @@ describe('Basic', function () {
     expect($('.fotorama__stage__frame', $fotorama).size()).toEqual(2);
     expect($('.fotorama__nav__frame', $fotorama).size()).toEqual(5);
   });
-  it('dimensions are correct', function () {
+  it('dimensions are correct', function (done) {
     var $body = $('body'),
         $stage = $('.fotorama__stage', $fotorama);
 
     waitsFor(function () {
       return $stage.height() > 0
-    }, 'Waits for autosize...', 100);
-
-    runs(function () {
+    }, function () {
+      expect($stage.height()).toBeGreaterThan(0);
       expect($stage.width()).toEqual($body.width());
       expect($stage.height()).toEqual(Math.round($body.width()/(700/467)));
+
+      done();
     });
+
+
   });
 });