+++ test/specs/events-n-methods.js
@@ -8,6 +8,16 @@ document.write(
     '</div>'
 );
 
+function waitsFor (test, fn) {
+  if (test()) {
+    fn();
+  } else {
+    setTimeout(function () {
+      waitsFor(test, fn);
+    }, 10);
+  }
+}
+
 describe('Events', function () {
   var $fotorama, fotorama, binded, e = {};
 
@@ -57,29 +67,29 @@ describe('Events', function () {
     expect(e.showend.i).toBe(1);
   });
 
-  it('fotorama:load & .show(\'>\')', function () {
+  it('fotorama:load & .show(\'>\')', function (done) {
     waitsFor(function () {
       return e.load && e.load.i === 2;
-    }, 'fotorama:load should be called 2 times', 1000);
+    }, function () {
+      fotorama.show('>');
+      expect(e.show.i).toBe(2);
+      expect(fotorama.activeIndex).toBe(1);
+      // auto event, no touch
+      expect(e.show.extra.user).toBeUndefined();
+      expect(e.show.extra.time).toBe(fotorama.options.transitionduration);
 
-    fotorama.show('>');
-    expect(e.show.i).toBe(2);
-    expect(fotorama.activeIndex).toBe(1);
-    // auto event, no touch
-    expect(e.show.extra.user).toBeUndefined();
-    expect(e.show.extra.time).toBe(fotorama.options.transitionduration);
+      waitsFor(function () {
+        return e.load.i === 3;
+      }, function () {
+        expect(e.load.extra.index).toBe(2);
+        expect(e.showend.i).toBe(2);
 
-    waitsFor(function () {
-      return e.load.i === 3;
-    }, 'fotorama:load should be called 3 times after .show()', 1000);
-
-    runs(function () {
-      expect(e.load.extra.index).toBe(2);
-      expect(e.showend.i).toBe(2);
+        done();
+      });
     });
   });
 
-  it('<, <<, >>, fotorama:error', function () {
+  it('<, <<, >>, fotorama:error', function (done) {
     fotorama.show('<'); // e.show.i = 3
     expect(fotorama.activeIndex).toBe(0);
     fotorama.show('>'); // e.show.i = 4
@@ -87,14 +97,12 @@ describe('Events', function () {
     fotorama.show('<<'); // e.show.i = 5
     expect(fotorama.activeIndex).toBe(0);
 
-    waitsFor(function () {
-      return e.error && e.error.i === 1;
-    }, 'fotorama:error', 1000);
-
     fotorama.show('>>'); // e.show.i = 6; e.showend.i = 3
     expect(fotorama.activeIndex).toBe(4);
 
-    runs(function () {
+    waitsFor(function () {
+      return e.error && e.error.i === 1;
+    }, function () {
       expect(e.show.i).toBe(6);
       expect(e.showend.i).toBe(3);
 
@@ -106,10 +114,12 @@ describe('Events', function () {
       // error frame was removed
       expect(fotorama.size).toBe(4);
       expect(fotorama.activeIndex).toBe(3);
+
+      done();
     });
   });
 
-  it('.startAutoplay(interval)', function () {
+  it('.startAutoplay(interval)', function (done) {
     expect(e.startautoplay).toBeUndefined();
     fotorama.startAutoplay(500);
     expect(e.startautoplay.i).toBe(1);
@@ -119,14 +129,19 @@ describe('Events', function () {
 
     waitsFor(function () {
       return fotorama.activeIndex === 1;
-    }, 'Autoplay rewinds to the start and continue', 500 * 2 + 333 + 100);
-
-    runs(function () {
+    }, function () {
       expect(e.show.i).toBe(8);
+      done();
     });
+
+
+
+    //runs(function () {
+
+    //});
   });
 
-  it('.stopAutoplay()', function () {
+  it('.stopAutoplay()', function (done) {
     expect(e.stopautoplay).toBeUndefined();
     fotorama.stopAutoplay();
     expect(e.startautoplay.i).toBe(1);
@@ -135,23 +150,21 @@ describe('Events', function () {
 
     var checked, index = fotorama.activeIndex;
 
-    runs(function () {
-      setTimeout(function () {
-        checked = true;
-      }, 600);
-    });
+    setTimeout(function () {
+      checked = true;
+    }, 600);
 
     waitsFor(function () {
       return checked;
-    }, 'Just wait...', 700);
-
-    runs(function () {
+    }, function () {
       // it is really stopped
       expect(fotorama.activeIndex).toBe(index);
+
+      done();
     });
   });
 
-  it('fullscreen events', function () {
+  it('fullscreen events', function (done) {
     fotorama.requestFullScreen();
     expect(e.fullscreenenter).toBeUndefined();
 
@@ -163,14 +176,20 @@ describe('Events', function () {
 
     waitsFor(function () {
       return fotorama.fullScreen;
-    }, 'Waiting for fotorama.fullScreen', 10);
-
-    runs(function () {
+    }, function () {
       fotorama.cancelFullScreen();
       expect(e.fullscreenenter.i).toBe(1);
       expect(e.fullscreenexit.i).toBe(1);
       expect(fotorama.fullScreen).toBe(false);
+
+      done();
     });
+
+
+
+    //runs(function () {
+
+    //});
 
   });