+++ test/specs/hash-n-show.js
@@ -8,6 +8,16 @@ document.write(
     '</div>'
 );
 
+function waitsFor (test, fn) {
+  if (test()) {
+    fn();
+  } else {
+    setTimeout(function () {
+      waitsFor(test, fn);
+    }, 10);
+  }
+}
+
 describe('Hash', function () {
   var $fotorama, fotorama;
 
@@ -20,15 +30,14 @@ describe('Hash', function () {
     expect(location.hash).toBe('');
   });
 
-  it('hash is a number', function () {
+  it('hash is a number', function (done) {
     fotorama.show('>');
 
     waitsFor(function () {
       return location.hash === '#2';
-    }, 'wait for a new hash', 500);
-
-    runs(function () {
+    }, function () {
       expect(location.hash).toBe('#' + (fotorama.activeIndex + 1));
+      done();
     });
   });